refactoring, and found an issue in the module loading
