+++ dds_tests/test_mod/os.py
@@ -1,0 +1,2 @@
+def test_function():
+    return "test_function"
